pyproject.toml
[tool.poetry]
name = "fastapitest"
version = "1.1.0"
description = ""
authors = ["It Soda <nikshchegilskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mysqlclient = "2.1.1"
fastapi = {extras = ["sqlalchemy"], version = "^0.104.1"}
python-dotenv = "^1.0.0"
alembic = "^1.13.0"
uvicorn = "^0.24.0.post1"
faker = "^20.1.0"
gunicorn = "^21.2.0"
pytest = "^7.4.3"
httpx = "^0.25.2"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
flake8 = "5.0.4"
isort = "5.10.1"
black = "23.10.1"
coverage = "7.3.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
